fix: refactor the method (#92)
