+++ accessors/spanner/spanner_accessor_test.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"os"
 	"testing"
+	"time"
 
 	"cloud.google.com/go/spanner"
 	"cloud.google.com/go/spanner/admin/database/apiv1/databasepb"
@@ -25,7 +26,6 @@ import (
 	spanneradmin "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/spanner/admin"
 	spannerclient "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/spanner/client"
 	spinstanceadmin "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/spanner/instanceadmin"
-	"github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/spanner/utils"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/constants"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/internal"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/logger"
@@ -33,6 +33,7 @@ import (
 	"github.com/googleapis/gax-go/v2"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
+	"golang.org/x/exp/rand"
 	"google.golang.org/api/iterator"
 )
 
@@ -49,6 +50,118 @@ const InterleaveDepthExpectError = "can't build CreateDatabaseRequest: can't cre
 const ColumnKeyPerTableExpectError = "can't build CreateDatabaseRequest: can't create/update database: can't create database: can't build CreateDatabaseRequest: rpc error: code = InvalidArgument desc = Table cart_extended has too many keys (17); the limit is 16."
 const TableNameExpectError = "can't build CreateDatabaseRequest: can't create/update database: can't create database: can't build CreateDatabaseRequest: rpc error: code = InvalidArgument desc = table name not valid: CustomerOrderTransactionHistoryRecords2023ForAnalysisAndArchivingIncludingSensitiveDataAndSecureProcessingProceduressrdfgdnhydbtsfvfs."
 const ColumnNameExpectError = "can't build CreateDatabaseRequest: can't create/update database: can't create database: can't build CreateDatabaseRequest: rpc error: code = InvalidArgument desc = Column name not valid: large_column.CustomerOrderTransactionHistoryRecords2023ForAnalysisAndArchivingIncludingSensitiveDataAndSecureProcessingProceduressrdfgdnhydbtsfvfs."
+
+// GenerateRandomString generates a random string of a specified length.
+func GenerateRandomString(length int) string {
+	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	var seededRand = rand.New(rand.NewSource(uint64(time.Now().UnixNano())))
+
+	randomString := make([]byte, length)
+	for i := range randomString {
+		randomString[i] = charset[seededRand.Intn(len(charset))]
+	}
+	return string(randomString)
+}
+
+// GenerateColumnDefsForTable generates an array of column definitions for a table
+// based on the specified length.
+func GenerateColumnDefsForTable(count int) map[string]ddl.ColumnDef {
+	colums := make(map[string]ddl.ColumnDef)
+	for i := 1; i <= count; i++ {
+		colName := fmt.Sprintf("col%d", i)
+		colId := fmt.Sprintf("c%d", i)
+		colums[colId] = ddl.ColumnDef{Name: colName, Id: colId, T: ddl.Type{Name: ddl.Int64}}
+	}
+	return colums
+}
+
+// GenerateColIds generates an array of column ids for a table
+// based on the specified length.
+func GenerateColIds(count int) []string {
+	var colIds []string
+	for i := 1; i <= count; i++ {
+		colId := fmt.Sprintf("c%d", i)
+		colIds = append(colIds, colId)
+	}
+	return colIds
+}
+
+// GeneratePrimaryColIds generates an array of primary columns ids for a table
+// based on the specified length.
+func GeneratePrimaryColIds(count int) []ddl.IndexKey {
+	var primaryKeys []ddl.IndexKey
+	for i := 1; i <= count; i++ {
+		colId := fmt.Sprintf("c%d", i)
+		primaryKeys = append(primaryKeys, ddl.IndexKey{ColId: colId})
+	}
+	return primaryKeys
+}
+
+// GenerateSpSchema generates a schema consisting of a specified number of tables.
+// Each table in the schema is defined by unique properties including identifiers,
+// primary keys, columns, and foreign keys, which are set based on the
+// iteration index and relationships with other tables.
+func GenerateSpSchema(count int) map[string]ddl.CreateTable {
+	spschema := make(map[string]ddl.CreateTable)
+	for i := 1; i <= count; i++ {
+		tableId := fmt.Sprintf("t%d", i)
+		tableName := fmt.Sprintf("table%d", i)
+		referTableId := fmt.Sprintf("t%d", i-1)
+		spschema[tableId] = ddl.CreateTable{
+			Name:        "table1",
+			Id:          tableName,
+			PrimaryKeys: GeneratePrimaryColIds(i),
+			ColIds:      GenerateColIds(i + 1),
+			ColDefs:     GenerateColumnDefsForTable(i + 1),
+			ForeignKeys: GenerateForeignKeys(i-1, referTableId),
+		}
+	}
+
+	return spschema
+}
+
+// GenerateForeignKeys generates an array of foreign keys for a table
+// based on the specified length.
+func GenerateForeignKeys(count int, referTableId string) []ddl.Foreignkey {
+	if count != 0 {
+		var colIds []string
+		var referColumnIds []string
+		for i := 1; i <= count; i++ {
+			colId := fmt.Sprintf("c%d", i)
+			colIds = append(colIds, colId)
+			referColumnIds = append(referColumnIds, colId)
+		}
+		fname := fmt.Sprintf("level%d_ibfk_1", count)
+		return []ddl.Foreignkey{{
+			Name:           fname,
+			ColIds:         colIds,
+			ReferColumnIds: referColumnIds,
+			ReferTableId:   referTableId,
+			Id:             GenerateRandomString(2),
+			OnDelete:       "NO ACTION",
+			OnUpdate:       "NO ACTION",
+		}}
+	} else {
+		return nil
+	}
+
+}
+
+// GenerateTables generates an array of tables
+// based on the specified length.
+func GenerateTables(count int) ddl.Schema {
+	tables := make(ddl.Schema)
+
+	for i := 1; i <= count; i++ {
+		tableName := fmt.Sprintf("table%d", i)
+		tableId := fmt.Sprintf("t%d", i)
+		tables[tableId] = ddl.CreateTable{Name: tableName, Id: tableId, PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}}, ColIds: []string{"c1"},
+			ColDefs: map[string]ddl.ColumnDef{
+				"c1": {Name: "col1", Id: "c1", T: ddl.Type{Name: ddl.Int64}},
+			}}
+	}
+	return tables
+}
 
 func init() {
 	logger.Log = zap.NewNop()
@@ -500,7 +613,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			expectedErrorMsg: TablePerDbExpectError,
 			expectError:      true,
 			dialect:          "google_standard_sql",
-			SpSchema:         utils.GenerateTables(5005),
+			SpSchema:         GenerateTables(5005),
 			migrationType:    "dataflow",
 		},
 		{
@@ -515,7 +628,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			expectedErrorMsg: "",
 			expectError:      false,
 			dialect:          "google_standard_sql",
-			SpSchema:         utils.GenerateTables(5000),
+			SpSchema:         GenerateTables(5000),
 			migrationType:    "dataflow",
 		},
 		{
@@ -533,8 +646,8 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 					Name:        "table 1",
 					Id:          "t1",
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
-					ColIds:      utils.GenerateColIds(1030),
-					ColDefs:     utils.GenerateColumnDefsForTable(1030),
+					ColIds:      GenerateColIds(1030),
+					ColDefs:     GenerateColumnDefsForTable(1030),
 				},
 			},
 			migrationType: "dataflow",
@@ -556,8 +669,8 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 					Name:        "table 1",
 					Id:          "t1",
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
-					ColIds:      utils.GenerateColIds(1024),
-					ColDefs:     utils.GenerateColumnDefsForTable(1024),
+					ColIds:      GenerateColIds(1024),
+					ColDefs:     GenerateColumnDefsForTable(1024),
 				},
 			},
 			migrationType: "dataflow",
@@ -574,7 +687,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			expectedErrorMsg: "",
 			expectError:      false,
 			dialect:          "google_standard_sql",
-			SpSchema:         utils.GenerateSpSchema(7),
+			SpSchema:         GenerateSpSchema(7),
 			migrationType:    "dataflow",
 		},
 		{
@@ -587,7 +700,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			expectedErrorMsg: InterleaveDepthExpectError,
 			expectError:      true,
 			dialect:          "google_standard_sql",
-			SpSchema:         utils.GenerateSpSchema(8),
+			SpSchema:         GenerateSpSchema(8),
 			migrationType:    "dataflow",
 		},
 		{
@@ -606,9 +719,9 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 				"t1": {
 					Name:        "table 1",
 					Id:          "t1",
-					PrimaryKeys: utils.GeneratePrimaryColIds(16),
-					ColIds:      utils.GenerateColIds(18),
-					ColDefs:     utils.GenerateColumnDefsForTable(18),
+					PrimaryKeys: GeneratePrimaryColIds(16),
+					ColIds:      GenerateColIds(18),
+					ColDefs:     GenerateColumnDefsForTable(18),
 				},
 			},
 			migrationType: "dataflow",
@@ -627,9 +740,9 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 				"t1": {
 					Name:        "table 1",
 					Id:          "t1",
-					PrimaryKeys: utils.GeneratePrimaryColIds(17),
-					ColIds:      utils.GenerateColIds(18),
-					ColDefs:     utils.GenerateColumnDefsForTable(18),
+					PrimaryKeys: GeneratePrimaryColIds(17),
+					ColIds:      GenerateColIds(18),
+					ColDefs:     GenerateColumnDefsForTable(18),
 				},
 			},
 			migrationType: "dataflow",
@@ -647,7 +760,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			dialect:          "google_standard_sql",
 			SpSchema: map[string]ddl.CreateTable{
 				"t1": {
-					Name:        utils.GenerateRandomString(130),
+					Name:        GenerateRandomString(130),
 					Id:          "t1",
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
 					ColIds:      []string{"c1"},
@@ -672,7 +785,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 			dialect:          "google_standard_sql",
 			SpSchema: map[string]ddl.CreateTable{
 				"t1": {
-					Name:        utils.GenerateRandomString(128),
+					Name:        GenerateRandomString(128),
 					Id:          "t1",
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
 					ColIds:      []string{"c1"},
@@ -700,7 +813,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
 					ColIds:      []string{"c1"},
 					ColDefs: map[string]ddl.ColumnDef{
-						"c1": {Name: utils.GenerateRandomString(130), Id: "c1", T: ddl.Type{Name: ddl.Int64}},
+						"c1": {Name: GenerateRandomString(130), Id: "c1", T: ddl.Type{Name: ddl.Int64}},
 					},
 				},
 			},
@@ -726,7 +839,7 @@ func TestSpannerAccessorImpl_CreateDatabase_exceeds_and_hit_limits(t *testing.T)
 					PrimaryKeys: []ddl.IndexKey{{ColId: "c1"}},
 					ColIds:      []string{"c1"},
 					ColDefs: map[string]ddl.ColumnDef{
-						"c1": {Name: utils.GenerateRandomString(128), Id: "c1", T: ddl.Type{Name: ddl.Int64}},
+						"c1": {Name: GenerateRandomString(128), Id: "c1", T: ddl.Type{Name: ddl.Int64}},
 					},
 				},
 			},
